Add getWeights method to IClassifier
During a BloomParamOpt routine, we optimize the hash allocation strategy
of a Bloom filter designed to represent the features of our data. To do
this, we need to know the weights that have been assigned to each of the
features, so that we can allocate them based on some notion of
"importance".

Adding a getWeights method to the IClassifier interface works mainly
because the IClassifier family knows exactly how to parse the
feature/feature map/weights files. Although it is a bit of a hack to
initialize an IClassifier object simply for this purpose, this method is
something that arguably should be there anyway.

If necessary we can later transition away from using the IClassifier
objects to parse these files, but for now it will do and the method
should be there anyway.