Update IClassifier interface and derived classes to use getWeights
IClassifiers need a reliable interface-based method to get the weight
vector. We add the `getWeights()` method in this commit. The big problem
here is that `getWeights()` must return `Hashtable[]`, since it must
allow for multiclass classifiers. Because of type erasure, Java runtime
is not able to differentiate between `Hashtable[]` and `Hashtable`. Thus
there is a single method that returns `Hashtable[]`.

Second, what we really want is a `Hashtable<String,Double>[]`, but again
due to type erasure, this is not possible, since we cannot ensure type
safety of the type parameters in Java. Life is pain.