Implements tool-name search next to indexing as label-getter
Currently, VertexProcessor can only obtain a label attribute by index in
the list. That is, if obtain the gender information of a communicant by
looking in the attribute list. However, if we want to look for the
gender attribute that is given by, say, dataset version 8 (rather than,
say, version 2), then we could look through the metadata in the list.

This commit introduces such searching abilities.